renames cmd package to have a "cmd" prefix
