+++ pkg/cmdinstall/install.go
@@ -1,4 +1,4 @@
-package install
+package cmdinstall
 
 import (
 	"flag"
