+++ pkg/cmdupdate/update.go
@@ -1,4 +1,4 @@
-package update
+package cmdupdate
 
 import (
 	"flag"
